DIGG-58: More refactoring
