+++ pages/search/index.ts
@@ -1,3 +1,3 @@
-import { SearchContentPage } from "../../components/pages/SearchContentPage";
+import { SearchContentPage } from "@/components/content/Search/SearchPage/SearchContentPage";
 
 export default SearchContentPage;
